@@ -1,13 +1,15 @@
-import { createRouter, createWebHashHistory } from "vue-router";
-
-const View1 = () => import("../views/link1.vue");
-const routes = [{
-    path: "/",
-    component: View1
-}];
+import { createRouter, createWebHistory, createWebHashHistory } from 'vue-router'
 
 const router = createRouter({
-    history: createWebHashHistory(),
-    routes
-})
+  history: createWebHashHistory(import.meta.env.BASE_URL),
+  routes: [
+    {
+      path: "/exper",
+      name: "jspsych",
+      component: () => import("@/views/exper/default.vue"),
+      alias: "/"
+    }
+  ]
+});
+
 export default router
